Create demo_class_function.php
just a function of a class from one project, the purpose of which is to save data from any table in the database in a separate table or jsion file according to the parameters specified in the parameters without passing unnecessary parameters.
It's taken out of context, but the goal is just to show the code.